Refactor to break transactions hooks out
