+++ cardstack/src/constants.ts
@@ -1,1 +1,2 @@
 export const ENABLE_PAYMENTS = true;
+export const TRANSACTION_PAGE_SIZE = 100;
