+++ cardstack/src/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './transaction-confirmation/use-transaction-confirmation';
 export * from './transaction-confirmation/use-is-message-request';
+export * from './transactions';
