+++ cardstack/src/services/index.ts
@@ -1,6 +1,6 @@
 export * from './gnosis-service';
-export * from './transaction-service';
 export * from './decoding-service';
 export * from './historical-pricing-service';
 export * from './exchange-rate-service';
 export * from './firebase-cloud-functions';
+export * from './transaction-mapping-service';
