+++ cardstack/src/utils/index.ts
@@ -6,3 +6,4 @@ export * from './ganache-utils';
 export * from './cardpay-utils';
 export * from './async-utils';
 export * from './date-utils';
+export * from './sorting-utils';
