+++ cardstack/src/utils/sorting-utils.ts
@@ -1,0 +1,6 @@
+export const sortByTime = (a: any, b: any) => {
+  const timeA = Number(a.timestamp || a.minedAt || a.createdAt);
+  const timeB = Number(b.timestamp || b.minedAt || b.createdAt);
+
+  return timeB - timeA;
+};
